Change structure of groups.
A group is now a Magma with a proof that the magma is a group. I tried 
to make it so there are still wrappers for all the old types, i.e., we 
can simply use Groups as we did before. From that perspective the only 
difference is that the ULevel is now an implicit argument. Similarly for 
Subgroups.

Also I moved towards using the HOTT library instead of our own utility 
functions.